[plugin] split files for the two RHs
